+++ src/Geopilot.Frontend/src/components/adminGrid/AdminGrid.tsx
@@ -59,55 +59,51 @@ export const AdminGrid: FC<AdminGridProps> = ({ addLabel, data, columns, onSave,
 
       if (isInEditMode) {
         return [
-          <GridActionsCellItem
-            key="save"
-            icon={
-              <Tooltip title={t("save")}>
-                <SaveOutlinedIcon />
-              </Tooltip>
-            }
-            label={t("save")}
-            onClick={handleSaveClick(id)}
-            color="inherit"
-          />,
-          <GridActionsCellItem
-            key="cancel"
-            icon={
-              <Tooltip title={t("cancel")}>
-                <CancelOutlinedIcon />
-              </Tooltip>
-            }
-            label={t("cancel")}
-            onClick={handleCancelClick(id)}
-            color="inherit"
-          />,
+          <Tooltip title={t("save")} key={`save-${id}`}>
+            <span>
+              <GridActionsCellItem
+                icon={<SaveOutlinedIcon />}
+                label={t("save")}
+                onClick={handleSaveClick(id)}
+                color="inherit"
+              />
+            </span>
+          </Tooltip>,
+          <Tooltip title={t("cancel")} key={`cancel-${id}`}>
+            <span>
+              <GridActionsCellItem
+                icon={<CancelOutlinedIcon />}
+                label={t("cancel")}
+                onClick={handleCancelClick(id)}
+                color="inherit"
+              />
+            </span>
+          </Tooltip>,
         ];
       }
 
       return [
-        <GridActionsCellItem
-          key="edit"
-          icon={
-            <Tooltip title={t("edit")}>
-              <EditOutlinedIcon />
-            </Tooltip>
-          }
-          label={t("edit")}
-          onClick={handleEditClick(id)}
-          color="inherit"
-          disabled={editingRow !== undefined}
-        />,
-        <GridActionsCellItem
-          key="disconnect"
-          icon={
-            <Tooltip title={t("disconnect")}>
-              <LinkOffIcon />
-            </Tooltip>
-          }
-          label={t("disconnect")}
-          onClick={handleDisconnectClick(id)}
-          color="error"
-        />,
+        <Tooltip title={t("edit")} key={`edit-${id}`}>
+          <span>
+            <GridActionsCellItem
+              icon={<EditOutlinedIcon />}
+              label={t("edit")}
+              onClick={handleEditClick(id)}
+              color="inherit"
+              disabled={editingRow !== undefined}
+            />
+          </span>
+        </Tooltip>,
+        <Tooltip title={t("disconnect")} key={`disconnect-${id}`}>
+          <span>
+            <GridActionsCellItem
+              icon={<LinkOffIcon />}
+              label={t("disconnect")}
+              onClick={handleDisconnectClick(id)}
+              color="error"
+            />
+          </span>
+        </Tooltip>,
       ];
     },
   };