+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Rename _Abgabe_ to _Lieferung_.
 - Refactored delivery overview to use only [MUI](https://mui.com/material-ui/) components.
 - Use Typescript for new components.
+- STAC browser now opens in a new tab.
 
 ## v1.0.93 - 2024-05-14